give more info on Cholesky failure and "fix" singular jacobian cols
I was getting NaN values in the parameters after the Cholesky decomposition
failed. To some extent, the issue arises in the a* phase, since that involves
dividing by values from the jacobian itself, which in this case were zero. The
root of the issue then is actually the jacobian, which shouldn't have a column
full of zeros in it, so I introduced some code to set the first element of a
column of zeros to 1.0 to combat this.