return a default spectro Output if there's an error in sic::freqs
this might cause trouble somewhere else, but at least it's a start. I think the
length checking elsewhere should salvage it, but there might be an issue in
particular if both directions of the num jac step go bad. at that point, I'll
probably need to pass the correct length around somehow so I can return a vec of
zeros of that size