the first loop here accounts for /= NU above, so start at 0
moving the i += 1 after the break was the actual fix for the issue I was seeing,
not both of them at once. for example, if we enter this loop once, we have 1
factor of 1/NU and then multiply in the loop by NU, restoring the original
lambda, so we would want to do pow(0). only if the first loop succeeds do we
need to start dividing off additional factors of NU.